BUG/MINOR: stream-int: also consider ENOTCONN in addition to EAGAIN
A failed send() may return ENOTCONN when the connection is not yet established.
On Linux, we generally see EAGAIN but on OpenBSD we clearly have ENOTCONN, so
let's ensure we poll for write when we encounter this error.
(cherry picked from commit 0ea0cf606e1da866b1c1e1b25dbe3472ccaaa6d8)